Huge refactor: Split wire-api crate and extract client-side game engine into crate
wire-api crate is now split into two crates and the proto types are built (kind of redundantly) in the client and server crates. This is awesome and allows me to implement traits on each type in a way that's specific to whether the trait (e.g. From<>) is being used by the client or server.

Moving the client-side game engine logic out of the client binary crate allows me to use the same client-side game engine to power a different UI/frontend in the future.